src/vectorDB/embeddings.py
from langchain_google_genai import GoogleGenerativeAIEmbeddings
from dotenv import load_dotenv
import os
load_dotenv()

embeddings = GoogleGenerativeAIEmbeddings(model=os.getenv("GOOGLE_EMBEDDING_MODEL"),
                                          google_api_key=os.getenv('GOOGLE_API_KEY'))
